pyproject.toml
[build-system]
requires = ["setuptools>=74"]
build-backend = "setuptools.build_meta"

[project]
name = "georeel"
version = "1.2.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "gpxpy>=1.6.2",
    "numpy>=2.4.4",
    "pillow>=12.2.0",
    "pyside6>=6.11.0",
    "scipy>=1.17.1",
    "srtm-py>=0.3.7",
]

[project.scripts]
georeel = "georeel.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
georeel = ["assets/*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--tb=short -q"

[tool.coverage.run]
source = ["georeel"]
omit = ["*/ui/*", "*/blender_scripts/*"]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-cov>=7.1.0",
]
